Next

+++ iron_bank_app/models.py
@@ -4,7 +4,7 @@ from django.db import models
 TYPE_CHOICES = [('d', 'deposit'),('w', 'withdrawal')]
 
 class AccountNumber(models.Model):
-    balance = models.IntegerField()
+    balance = models.FloatField()
     user = models.ForeignKey(User)
     nickname = models.CharField(max_length=30)
     created_time = models.DateTimeField(auto_now_add=True)
@@ -15,7 +15,7 @@ class AccountNumber(models.Model):
 class Transaction(models.Model):
     account = models.ForeignKey(AccountNumber)
     trans_type = models.CharField(max_length=1, choices=TYPE_CHOICES)
-    amount = models.IntegerField()
+    amount = models.FloatField()
     description = models.CharField(max_length=50)
     time_created = models.DateTimeField(auto_now_add=True)
 
@@ -24,7 +24,7 @@ class Transaction(models.Model):
 
 class Transfer(models.Model):
     account = models.ForeignKey(AccountNumber)
-    amount = models.IntegerField()
+    amount = models.FloatField()
     transfer_time = models.DateTimeField(auto_now_add=True)
 
     class Meta:
